Time: 39 ms (8.59%), Space: 48.3 MB (5.15%) - LeetHub

+++ 0322-coin-change/0322-coin-change.java
@@ -1,5 +1,5 @@
 class Solution {
-    private long  minCoins(int idx,int[][]dp,int[]coins,int amount){
+    private long  minCoins(int idx,long[][]dp,int[]coins,int amount){
         if(idx==coins.length ){
             if(amount==0) return 0;//no more coins needed;
             else return Integer.MAX_VALUE;// not avalid combination
@@ -7,14 +7,14 @@ class Solution {
 
         if(dp[idx][amount]!=-1) return dp[idx][amount];
         long skip=minCoins(idx+1,dp,coins,amount);
-        if(amount-coins[idx]<0)return dp[idx][amount]=(int)skip;
+        if(amount-coins[idx]<0)return dp[idx][amount]=skip;
         long take=1+minCoins(idx,dp,coins,amount-coins[idx]);
 
-        return dp[idx][amount]=(int)Math.min(skip,take);
+        return dp[idx][amount]=Math.min(skip,take);
     }
     public int coinChange(int[] coins, int amount) {
         if(amount==0) return 0;
-        int[][] dp=new int[coins.length][amount+1];
+        long[][] dp=new long[coins.length][amount+1];
         for(int i=0;i<dp.length;i++){
             for(int j=0;j<dp[0].length;j++){
                 dp[i][j]=-1;
